ThumbnailGenerator/requirements.txt
Pillow==3.0.0

